src/products/urls.py
from rest_framework.routers import DefaultRouter

from products.views import ProductViewSet


router = DefaultRouter()
router.register('', ProductViewSet, basename='product')

urlpatterns = []
urlpatterns += router.urls
